+++ src/app/service/terra-base.service.ts
@@ -9,6 +9,7 @@ import 'rxjs/add/operator/map';
 import { Observable } from 'rxjs';
 import { Exception } from './data/exception.interface';
 import {
+    isArray,
     isNull,
     isNullOrUndefined
 } from 'util';
@@ -273,9 +274,10 @@ export class TerraBaseService
 
     /**
      * @param {TerraBaseParameterInterface} params
+     * @param {boolean} arrayAsArray - Defines if an array search param should interpret and parsed as an array or not. Default is false.
      * @returns {URLSearchParams}
      */
-    protected createUrlSearchParams(params:TerraBaseParameterInterface):URLSearchParams
+    protected createUrlSearchParams(params:TerraBaseParameterInterface, arrayAsArray:boolean = false):URLSearchParams
     {
         let searchParams:URLSearchParams = new URLSearchParams('', new TerraQueryEncoder());
 
@@ -285,12 +287,32 @@ export class TerraBaseService
             {
                 if(!isNullOrUndefined(params[key]) && params[key] !== '')
                 {
-                    searchParams.set(key, params[key]);
+                    if(arrayAsArray && isArray(params[key]))
+                    {
+                        searchParams.appendAll(this.createArraySearchParams(key, params[key]));
+                    }
+                    else
+                    {
+                        searchParams.set(key, params[key]);
+                    }
                 }
+
             });
         }
 
         return searchParams;
+    }
+
+    private createArraySearchParams(key:string, params:Array<string>):URLSearchParams
+    {
+        let arraySearchParams:URLSearchParams = new URLSearchParams();
+
+        params.forEach((param:string) =>
+        {
+            arraySearchParams.append(key + '[]', param);
+        });
+
+        return arraySearchParams;
     }
 
     private getMissingUserPermissionAlertMessage(error:any):string