+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	* added the `inputButtonList` input to display buttons using the `TerraButtonInterface`
 	* added the `inputNoWordBreak` input to disable the auto word break if set to true
 * **unit testing** added packages for testing with jasmine framework and karma command line tool
+* **terra-base-service** add `arrayAsArray` (default `false`) parameter to `createUrlSearchParams` function. If set to `true` an array is parsed to an array parameter and not a concatenated string.
 
 ### Bug Fixes
 * **terra-button** "flagged" buttons are now also clickable in the yellow corner